+++ CHANGELOG.md
@@ -14,12 +14,17 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds `displaySize` prop to `Heading`
 - Adds `helperText` and `errorText` props to `Checkbox`
 - Adds `helperText` and `errorText` props to `Radio`
+- Adds `errorText`, `helperText`, `labelText`, `required` and `showLabel` props to `Select`
 
 ### Changes
 
 - Updates content in Style Guide `Typography`
 - Updates content in Style Guide `Breakpoints`
 - Changes `Link` story from `.tsx` to `.mdx`.
+
+### Deprecates
+
+- Deprecates `ariaLabel`, `helperTextId`, `isRequired` and `labelId` in `Select`
 
 ## 0.23.0